requirements.txt
streamlit
opencv-python
numpy
pillow
google-cloud-vision
fpdf
boto3
pandas
xlsxwriter
